clippy: fix -Dwarnings and document unsafe fns
Bring the kernel and hostlib under clippy -D warnings: mechanical
fixes from clippy --fix, site-level allows with reasons, # Safety
on every unsafe fn, and FdError instead of Result<_, ()>.
Keep kernel_tests dead_code allows (Q2) and parked shell ready (#66).

+++ src/fs_init.rs
@@ -66,8 +66,8 @@ fn attach_pseudo_dirs() {
 
 fn populate_devfs() {
     with(|v| {
-        let sz = (block_init::capacity_sectors() as u64)
-            .saturating_mul(block_init::logical_block_size() as u64);
+        let sz =
+            block_init::capacity_sectors().saturating_mul(block_init::logical_block_size() as u64);
         let _ = v.devfs_add_block(block_init::name().as_bytes(), sz);
         if virtio_blk_init::live() {
             let sz = virtio_blk_init::capacity_sectors()
